backend/requirements/base.txt
Django>=5.2,<5.3
djangorestframework>=3.17,<3.18
djangorestframework-simplejwt>=5.5,<5.6
django-cors-headers>=4.9,<5.0
django-filter>=25.2,<26.0
django-environ>=0.13,<0.14
django-unfold>=0.86,<0.87
celery>=5.6,<5.7
redis>=7.3,<7.4
